Compleet

+++ app/Http/Controllers/Dashbord/CategoryControllerr.php
@@ -2,17 +2,24 @@
 
 namespace App\Http\Controllers\Dashbord;
 
-use App\Http\Controllers\Controller;
-use App\Models\Category;
-use Illuminate\Http\Request;
 use DataTables;
+use App\Models\Setting;
+use App\Models\Category;
 use Illuminate\Support\Str;
+use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 
 
 
 
 class CategoryControllerr extends Controller
 {
+
+    protected $setting;
+    public function __construct(Setting $setting)
+    {
+        $this->setting = $setting;
+    }
     /**
      * Display a listing of the resource.
      */
@@ -28,19 +35,22 @@ class CategoryControllerr extends Controller
      */
     public function create()
     {
-
+        $this->authorize('viewAny', $this->setting);
         $categories = Category::whereNull('parent')->orwhere('parent', 0)->get();
         return view('dashbord.categories.add' , compact('categories'));
     }
     public function getUsersDatatable(){
         $data = Category::select('*')->with('parents');
         return Datatables::of($data)
-            ->addIndexColumn()
-            ->addColumn('action', function($row){
-                return $btn = '<a href="' . Route('dashbord.category.edit', $row->id) . '" class="adit btn btn-success btn-sm" ><i class="fa fa-edit"></i></a><a
-                id="deleteBtn" data-id="' . $row->id . '" class="edit btn btn-danger btn-sm" data-toggle="modal"
-                data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
-            })
+        ->addIndexColumn()
+        ->addColumn('action', function ($row) {
+
+            if (auth()->user()->can('viewAny', $this->setting)) {
+                return $btn = '
+                    <a href="' . Route('dashbord.category.edit', $row->id) . '"  class="edit btn btn-success btn-sm" ><i class="fa fa-edit"></i></a>
+                    <a id="deleteBtn" data-id="' . $row->id . '" class="edit btn btn-danger btn-sm"  data-toggle="modal" data-target="#deletemodal"><i class="fa fa-trash"></i></a>';
+            }
+        })
             ->addColumn('parent', function ($row) {
                 return ($row->parent ==  0) ? trans('words.main category') :   $row->parents->translate(app()->getLocale())->title;
             })
@@ -66,7 +76,7 @@ class CategoryControllerr extends Controller
      */
     public function store(Request $request )
     {
-
+        $this->authorize('viewAny', $this->setting);
        $category = Category::create($request->except('image','_token'));
        $category->update($request->except('image', '_token'));
         if ($request->file('image')) {
@@ -93,6 +103,7 @@ class CategoryControllerr extends Controller
      */
     public function edit(Category $category)
     {
+        $this->authorize('viewAny', $this->setting);
         $categories = Category::whereNull('parent')->orwhere('parent', 0)->get();
         return view ('dashbord.categories.edit', compact('category','categories'));
     }
@@ -102,7 +113,7 @@ class CategoryControllerr extends Controller
      */
     public function update(Request $request, Category $category)
     {
-
+        $this->authorize('viewAny', $this->setting);
         $category->update($request->except('image', '_token'));
         if ($request->file('image')) {
             $file = $request->file('image');
@@ -126,6 +137,7 @@ class CategoryControllerr extends Controller
 
     public function delete(Request $request){
 
+        $this->authorize('viewAny', $this->setting);
         if(is_numeric($request->id)) {
             Category::where('parent', $request->id)->delete();
             Category::where('id', $request->id)->delete();
